Fix invalid timeout for update function
